+++ src/app/page.tsx
@@ -7,5 +7,7 @@ import dynamic from "next/dynamic"
 const SwaggerUI = dynamic(() => import("swagger-ui-react"), { ssr: false })
 
 export default function Swagger() {
-	return <SwaggerUI url="/api/openapi.json" />
+	const url = `${process.env.NODE_ENV == "production" ? "/lol-scan-api" : ""}/api/openapi.json`
+
+	return <SwaggerUI url={url} />
 }