+++ src/server/trpc.ts
@@ -1,8 +1,8 @@
-import { TRPCError, initTRPC } from "@trpc/server"
+import { initTRPC } from "@trpc/server"
 import { OpenApiMeta } from "trpc-openapi"
 import { TRPCPanelMeta } from "trpc-panel"
 
-import { checkApiKey } from "@/lib/error"
+import { checkApiKey, checkRateLimit } from "@/lib/error"
 import rateLimit from "@/lib/rateLimit"
 
 type Context = {
@@ -15,22 +15,12 @@ const t = initTRPC.meta<Meta>().context<Context>().create()
 
 export const router = t.router
 
-const limit = rateLimit({
-	intervalInMs: 60 * 1000,
-	uniqueTokenPerInterval: 10,
-})
+const limiter = rateLimit()
 
 export const auth = t.middleware(async ({ ctx, next }) => {
 	checkApiKey(ctx.apiKey)
 
-	try {
-		await limit.check(100, "RATE_LIMIT_CACHE_TOKEN")
-	} catch {
-		throw new TRPCError({
-			code: "TOO_MANY_REQUESTS",
-			message: "Rate limit exceeded 100 requests per minute",
-		})
-	}
+	await checkRateLimit(limiter, 100)
 
 	return next({
 		ctx: {