+++ tests/app/swagger.spec.ts
@@ -1,0 +1,7 @@
+import test, { expect } from "@playwright/test"
+
+test("title_swaggerUi", async ({ page }) => {
+	await page.goto("http://localhost:3000")
+
+	await expect(page.getByText("lol-scan-api")).toBeVisible()
+})