+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 	Changelog
 </h1>
 
+### 1.1.4
+
+#### New features
+
+-   Swagger UI test
+
+#### Changes
+
+-   separate rate limit check
+-   Swagger UI URL
+-   Playwright config
+
 ### 1.1.3
 
 #### New features